# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Relation API",
        "version": "1.0"
    },
    "paths": {
        "/api/v0/licenses": {
            "get": {
                "tags": [
                    "Licenses Data"
                ],
                "summary": "Get Licenses",
                "operationId": "getLicenses",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Licenses"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "/api/v0/licenses?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": "10"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Post Licenses Data"
                ],
                "summary": "Post Licenses",
                "operationId": "postLicenses",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "license_name": {
                                        "type": "string",
                                        "example": "HACCP"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": "1013"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "License has been created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Licenses"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/v0/licenses/{id}": {
            "get": {
                "tags": [
                    "Licenses by Id Data"
                ],
                "summary": "Get License by Id",
                "operationId": "getLicenseId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the license",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "21"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Licenses"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "/api/v0/licenses?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": "10"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Update License Data"
                ],
                "summary": "Update License Data",
                "operationId": "updateLicense",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the license",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "21"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "license_name": {
                                        "type": "string",
                                        "example": "ISO 5001"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": "1018"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "License has been updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Licenses"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Delete license Data"
                ],
                "summary": "Delete License",
                "operationId": "deleteLicense",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the license",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "license has been deleted"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Licenses": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "license_name": {
                        "type": "string",
                        "example": "ISO 20001"
                    },
                    "product_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-05-13T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-05-13T00:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    }
}